+++ apps/common/handle/base_parse_table_handle.py
@@ -15,5 +15,5 @@ class BaseParseTableHandle(ABC):
         pass
 
     @abstractmethod
-    def handle(self, file, get_buffer):
+    def handle(self, file, get_buffer,save_image):
         pass
