+++ apps/common/handle/impl/table/excel_parse_table_handle.py
@@ -5,6 +5,7 @@ import logging
 from openpyxl import load_workbook
 
 from common.handle.base_parse_table_handle import BaseParseTableHandle
+from common.handle.impl.tools import xlsx_embed_cells_images
 
 max_kb = logging.getLogger("max_kb")
 
@@ -16,10 +17,15 @@ class ExcelSplitHandle(BaseParseTableHandle):
             return True
         return False
 
-    def handle(self, file, get_buffer):
+    def handle(self, file, get_buffer, save_image):
         buffer = get_buffer(file)
         try:
             wb = load_workbook(io.BytesIO(buffer))
+            try:
+                image_dict: dict = xlsx_embed_cells_images(io.BytesIO(buffer))
+                save_image([item for item in image_dict.values()])
+            except Exception as e:
+                image_dict = {}
             result = []
             for sheetname in wb.sheetnames:
                 paragraphs = []
@@ -35,7 +41,11 @@ class ExcelSplitHandle(BaseParseTableHandle):
                             continue
                         t = str(ti[i].value) if i < len(ti) else ""
                         title.append(t)
-                        t += (": " if t else "") + str(c.value)
+                        content = str(c.value)
+                        image = image_dict.get(content, None)
+                        if image is not None:
+                            content = f'![](/api/image/{image.id})'
+                        t += (": " if t else "") + content
                         l.append(t)
                     l = "; ".join(l)
                     if sheetname.lower().find("sheet") < 0:
